+++ src/ska/contrib/django/ska/settings.py
@@ -10,6 +10,9 @@ from __future__ import absolute_import
 - `USER_CREATE_CALLBACK` (str): User create callback (when user is created in auth backend).
 - `USER_INFO_CALLBACK` (str): User info callback.
 - `REDIRECT_AFTER_LOGIN` (str): Redirect after login.
+- `DB_STORE_SIGNATURES` (bool): If set to True, signatures are stored in the database.
+- `DB_PERFORM_SIGNATURE_CHECK` (bool): If set to True, an extra check is fired on whether the token has
+  already been used or not.
 """
 
 __title__ = 'ska.contrib.django.ska.settings'
@@ -18,7 +21,8 @@ __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'UNAUTHORISED_REQUEST_ERROR_MESSAGE', 'UNAUTHORISED_REQUEST_ERROR_TEMPLATE', 'AUTH_USER', 'SECRET_KEY',
-    'USER_GET_CALLBACK', 'USER_CREATE_CALLBACK', 'USER_INFO_CALLBACK', 'REDIRECT_AFTER_LOGIN'
+    'USER_GET_CALLBACK', 'USER_CREATE_CALLBACK', 'USER_INFO_CALLBACK', 'REDIRECT_AFTER_LOGIN',
+    'DB_STORE_SIGNATURES', 'DB_PERFORM_SIGNATURE_CHECK'
 )
 
 from django.conf import settings
@@ -39,3 +43,6 @@ USER_GET_CALLBACK = get_setting('USER_GET_CALLBACK')
 USER_CREATE_CALLBACK = get_setting('USER_CREATE_CALLBACK')
 USER_INFO_CALLBACK = get_setting('USER_INFO_CALLBACK')
 REDIRECT_AFTER_LOGIN = get_setting('REDIRECT_AFTER_LOGIN')
+
+DB_STORE_SIGNATURES = get_setting('DB_STORE_SIGNATURES')
+DB_PERFORM_SIGNATURE_CHECK = get_setting('DB_PERFORM_SIGNATURE_CHECK')